abstract config loader
