+++ src/ton_engine/data_source/postgres_source.py
@@ -1,31 +1,8 @@
-from sqlalchemy import create_engine, text
+from ton_engine.data_source.sqlalchemy_source import SqlAlchemyDataSource
 
 
-class PostgresDataSource:
+class PostgresDataSource(SqlAlchemyDataSource):
     def __init__(self, host: str, port: int, user: str, password: str, database: str):
-        self.host = host
-        self.port = port
-        self.user = user
-        self.password = password
-        self.database = database
-
-        self._engine = None
-
-    def setup(self):
-        """Run setup."""
-        self._engine = create_engine(
-            f'postgresql+psycopg://{self.user}:{self.password}@{self.host}:{self.port}/{self.database}'
+        super().__init__(
+            url=f'postgresql+psycopg://{user}:{password}@{host}:{port}/{database}',
         )
-
-    def run_query(self, query: str) -> list[tuple]:
-        """Run query on database and return its result."""
-        sql = text(query)
-
-        with self._engine.connect() as connection:
-            results = connection.execute(sql)
-
-        return list(results)
-
-    def teardown(self):
-        """Close connection."""
-        self._engine.dispose()
